add api restriction to attendance

+++ app/controllers/api/v1/attendances_controller.rb
@@ -6,36 +6,58 @@ class Api::V1::AttendancesController < ApplicationController
 
 
   def index
-    @attendances = Attendance.all
+
+    if current_member
+      @attendances = Attendance.all
+    else
+      head(:unauthorized)
+    end
   end
 
 
   def create
-    @reservation = Attendance.new(
-        event_id: params[:event_id],
-        member_id: params[:member_id],
-        time: params[:time]
-    )
 
-    if @reservation.save
-      @event = Event.find(params[:event_id])
-      @eventAttendance = @event.attendances
-      render :create, status: :created
+    if current_member
+
+      @reservation = Attendance.new(
+          event_id: params[:event_id],
+          member_id: params[:member_id],
+          time: params[:time]
+      )
+
+      if @reservation.save
+        @event = Event.find(params[:event_id])
+        @eventAttendance = @event.attendances
+        render :create, status: :created
+      else
+        head(:bad_request)
+      end
+
     else
-      head(:bad_request)
+      head(:unauthorized)
     end
+
   end
 
   def destroy
-    @reservation = Attendance.find(params[:id])
 
-    if @reservation.destroy
-      head(:ok)
+    if current_member
+
+      @reservation = Attendance.find(params[:id])
+
+      if @reservation.destroy
+        head(:ok)
+      else
+        head(:bad_request)
+      end
+
     else
-      head(:bad_request)
+      head(:unauthorized)
     end
 
   end
 
-
 end
+
+
+
